// WADNR.Web/src/app/shared/generated/model/fund-source-image-grid-row.ts
export class FundSourceImageGridRow { 
    FundSourceImageID?: number;
    FundSourceID?: number;
    FileResourceID?: number;
    FileResourceGuid?: string;
    Caption?: string | null;
    Credit?: string | null;
    IsKeyPhoto?: boolean;
    CreatedDate?: string;
    OriginalFilename?: string | null;
    ContentLength?: number | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}