// WADNR.Web/src/app/shared/generated/model/fund-source-image-upsert-request.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class FundSourceImageUpsertRequest { 
    Caption?: string | null;
    Credit?: string | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface FundSourceImageUpsertRequestForm { 
    Caption?: FormControl<string>;
    Credit?: FormControl<string>;
}

export class FundSourceImageUpsertRequestFormControls { 
    public static Caption = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static Credit = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}